moved CollectionOperation.checkMedia to DeckPicker
