+++ AnkiDroid/src/test/java/com/ichi2/anki/DeckPickerCheckMediaTest.kt
@@ -1,5 +1,6 @@
 /*
  *  Copyright (c) 2020 David Allison <davidallisongithub@gmail.com>
+ *  Copyright (c) 2022 Divyansh Kushwaha <kushwaha.divyansh.dxn@gmail.com>
  *
  *  This program is free software; you can redistribute it and/or modify it under
  *  the terms of the GNU General Public License as published by the Free Software
@@ -13,21 +14,28 @@
  *  You should have received a copy of the GNU General Public License along with
  *  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.ichi2.libanki
 
+package com.ichi2.anki
+
+import android.content.Intent
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.ichi2.anki.CollectionManager.withCol
-import com.ichi2.anki.RobolectricTest
-import com.ichi2.async.checkMedia
 import net.ankiweb.rsdroid.BackendFactory
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.robolectric.Robolectric
 import java.util.concurrent.ExecutionException
 
 @RunWith(AndroidJUnit4::class)
-class CheckMediaTest : RobolectricTest() {
+class DeckPickerCheckMediaTest : RobolectricTest() {
+    private lateinit var mImpl: DeckPicker
+    override fun setUp() {
+        super.setUp()
+        mImpl = Robolectric.buildActivity(DeckPicker::class.java, Intent())
+            .create().start().resume().get()
+    }
+
     override fun useInMemoryDatabase(): Boolean {
         return false
     }
@@ -45,7 +53,7 @@ class CheckMediaTest : RobolectricTest() {
             col.media.db!!.queryScalar("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='meta';"),
             equalTo(0)
         )
-        withCol { checkMedia(this) }
+        mImpl.checkMedia()
         assertThat(
             col.media.db!!.queryScalar("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='meta';"),
             equalTo(1)
